Добавил добавление/удаление избранного, поменял значения кнопки на enum, перенес логику в компонент кнопки, небольшие правки
