+++ src/components/place-card/place-card.tsx
@@ -1,12 +1,10 @@
-import { AppRoute, AuthStatus } from '@src/const';
+import { AppRoute } from '@src/const';
 import type { Offer } from '@src/entities/offers';
 import type { HTMLProps } from 'react';
-import { Link, useNavigate } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 import BookmarkButton from '../bookmark-button/bookmark-button';
 import Rating from '../rating/rating';
 import { capitalizeFirstLetter } from '@src/utils/formatters';
-import { useAppSelector } from '@src/hooks/store-hooks';
-import { getAuthStatus } from '@src/store/slices/user-slice';
 
 type PlaceCard = HTMLProps<HTMLElement> & {
   place: Omit<Offer, 'city'>;
@@ -33,15 +31,6 @@ export default function PlaceCard(props: PlaceCard): JSX.Element {
     rating,
     previewImage,
   } = place;
-
-  const userStatus = useAppSelector(getAuthStatus);
-  const navigate = useNavigate();
-
-  const handleBookmarkClick = () => {
-    if (userStatus !== AuthStatus.Auth) {
-      navigate(AppRoute.Login);
-    }
-  };
 
   return (
     <article className={`${bemblock}__card place-card`} {...htmlProps}>
@@ -74,10 +63,10 @@ export default function PlaceCard(props: PlaceCard): JSX.Element {
 
           {/* Добавление в закладки */}
           <BookmarkButton
+            offerID={id}
             bemblock="place-card"
             isFavorite={isFavorite}
             iconSize={{ width: 18, height: 19 }}
-            onClick={handleBookmarkClick}
           />
         </div>
